move setters/getters into subclass
